requirements.txt
transformers>=4.41
datasets>=2.19
peft>=0.10
bitsandbytes>=0.43
accelerate>=0.29
scikit-learn
python-dotenv
gradio
matplotlib
seaborn